I got a window to show up!

+++ Source/DX11/DxDevice.cpp
@@ -32,7 +32,7 @@ DxDevice::DxDevice(DxInstance *inst, const DeviceParams *params)
 void DxDevice::CreateSpriteSet(
     bool streaming, uint32_t spriteCount,
     uint32_t spriteWidth, uint32_t spriteHeight,
-    const uint32_t **buffers,
+    const uint8_t **buffers,
     ISpriteSet **set)
 {
     if (!streaming) assert(buffers);
